src/Room.vue: buttons to join via App
This change adds 2 buttons at the bottom of the Room page which allow
users to join a Jitsi room via the Jitsi Desktop app and the Jitsi
Mobile app, respectively.

The change moves calculations of the coresponding join URLs into
`computed()` as the button URLs are required when the page loads.
Joining with the desktop app requires the URL to be prefixed with
`jitsi-meet://`, while the mobile app requires `http[s]://`.

The "Join using the Jitsi App" section was updated to provide help for
joining with the Jitsi app. The calculated join HTTPS URL is re-used
there; a user may fall back to pasting this URL into the Jitsi app
should the buttons not work for them.

Signed-off-by: Thilo Fromm <thilo.alexander@gmail.com>